Refactor file structure of IntegrationTests solution
