+++ Catalog/Catalog.IntegrationTests/Catalog.IntegrationTests.csproj
@@ -28,4 +28,8 @@
       <ProjectReference Include="..\Catalog.Infrastructure\Catalog.Infrastructure.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Catalog.Infrastructure\Database\" />
+    </ItemGroup>
+
 </Project>
